#275 Notify outbox service that a message is available for publishing
