+++ src/Samples/Sample.OutboxWebApi/Program.cs
@@ -33,6 +33,7 @@ var configuration = builder.Configuration;
 // doc:fragment:ExampleStartup
 builder.Services.AddSlimMessageBus(mbb =>
 {
+    mbb.PerMessageScopeEnabled(false);
     mbb
         .AddChildBus("Memory", mbb =>
         {
@@ -73,6 +74,7 @@ builder.Services.AddSlimMessageBus(mbb =>
         .AddOutboxUsingDbContext<CustomerContext>(opts =>
         {
             opts.PollBatchSize = 100;
+            opts.PollIdleSleep = TimeSpan.FromSeconds(10);
             opts.MessageCleanup.Interval = TimeSpan.FromSeconds(10);
             opts.MessageCleanup.Age = TimeSpan.FromMinutes(1);
             //opts.SqlSettings.TransactionIsolationLevel = System.Data.IsolationLevel.RepeatableRead;
