+++ src/SlimMessageBus.Host.Outbox/IOutboxNotificationService.cs
@@ -1,0 +1,9 @@
+﻿namespace SlimMessageBus.Host.Outbox.Services;
+
+public interface IOutboxNotificationService
+{
+    /// <summary>
+    /// Notify outbox service that a message is waiting to be published.
+    /// </summary>
+    void Notify();
+}
